validando dados do form e atualizando banco de dados

+++ Views/Ativo/tabela.php
@@ -5,11 +5,11 @@
         <th>Filial</th>
         <th>Setor</th>
         <th>Categoria</th>
-        <th>Data_Cadastro</th>
-        <th>Data_Aquisição</th>
-        <th>Vida_Util</th>
+        <th>Data cadastro</th>
+        <th>Data aquisição</th>
+        <th>Vida útil</th>
         <th>Condição</th>
-        <th>Estado_Ativo</th>
+        <th>Estado ativo</th>
         <th>Valor</th>
     </tr>
     <?php foreach ($ativos as $ativo) : ?>
@@ -36,7 +36,7 @@
                 <?php echo traduz_data_para_exibir($ativo->getDataAquisicao()); ?>
             </td>
             <td>
-                <?php echo $ativo->getVidaUtil(); ?>
+                <?php echo $ativo->getVidaUtil() . ' ano'; ?>
             </td>
             <td>
                 <?php echo traduz_condicao_para_exibir($ativo->getCondicao()); ?>
@@ -51,7 +51,7 @@
                 <a href="?rota=ativo&edit_id=<?php echo $ativo->getId(); ?>">
                     Editar
                 </a>
-                <a href="?rota=ativo&delete_id=<?php echo $ativo->getId(); ?>">
+                <a href="?rota=ativo&delete_id=<?php echo $ativo->getId(); ?>" onclick="return confirmarRemocao()">
                     Remover
                 </a>
             </td>
